Move convertOffersToSimple and SimpleOffer into offers utils
