+++ src/components/PlaceOfferSection/PlaceOfferContent/components/Diagram/helpers.ts
@@ -1,8 +1,7 @@
 import { chunk, clamp, concat, filter } from 'lodash'
 
 import { Loan } from '@banx/api/core'
-import { SimpleOffer } from '@banx/pages/BorrowPage/types'
-import { calcLoanBorrowedAmount } from '@banx/utils'
+import { SimpleOffer, calcLoanBorrowedAmount } from '@banx/utils'
 
 import { Mark } from './Diagram'
 import { MAX_BOUND_PERCENTAGE, MAX_GROUP_SIZE, MIN_BOUND_PERCENTAGE } from './constants'
