+++ src/pages/BorrowPage/index.ts
@@ -1,2 +1,1 @@
 export { BorrowPage } from './BorrowPage'
-export * from './helpers'
