+++ src/utils/offers/index.ts
@@ -1,34 +1,2 @@
-import { getMaxLoanValueFromBondOffer } from 'fbonds-core/lib/fbond-protocol/helpers'
-import { BondOfferV2, PairState } from 'fbonds-core/lib/fbond-protocol/types'
-
-import { Offer } from '@banx/api/core'
-
-export const сalculateLoansAmount = (offer: Offer) => {
-  const { fundsSolOrTokenBalance, currentSpotPrice } = offer
-
-  const loansAmount = fundsSolOrTokenBalance / currentSpotPrice
-
-  return loansAmount
-}
-
-export const calculateLoanValue = (offer: Offer) => {
-  return getMaxLoanValueFromBondOffer(offer as BondOfferV2)
-  // const { currentSpotPrice } = offer
-
-  // const loansAmount = сalculateLoansAmount(offer)
-  // const loanValue = currentSpotPrice * Math.min(loansAmount, 1)
-
-  // return loanValue
-}
-
-export const isOfferClosed = (pairState: string) => {
-  return (
-    pairState === PairState.PerpetualClosed ||
-    pairState === PairState.PerpetualBondingCurveClosed ||
-    pairState === PairState.PerpetualMigrated
-  )
-}
-
-//? Prevent orders wrong distibution on bulk borrow from same offer
-export const offerNeedsReservesOptimizationOnBorrow = (offer: Offer, loanValueSum: number) =>
-  loanValueSum <= (offer.bidSettlement + offer.buyOrdersQuantity > 0 ? offer.currentSpotPrice : 0)
+export * from './helpers'
+export * from './types'
