+++ src/utils/offers/types.ts
@@ -1,0 +1,6 @@
+export interface SimpleOffer {
+  id: string
+  publicKey: string
+  loanValue: number
+  hadoMarket: string
+}
